Code Signing/Removal of DMG Generation
Dropping support for creating DMG. Ran in to a code signing issue that requires parts of the app be signed before they would be bundled into the DMG. Rather than mount the dmg and code sign, opted to just drop the dmg. Could likely be added easily in the future.

Also added the code signing portions so mac will ignore some of the popups that were happening. If we were to deploy on say the app store, we would need to get an authentic cert that mac would recognize instead of a self signed cert.

To build this on your own mac, you will need to create a code signing key called "netlock" unless you modify the build script.